Move LocalEventTimelineItem into its own module
