+++ crates/matrix-sdk/src/room/timeline/event_item/local.rs
@@ -1,0 +1,39 @@
+use ruma::{EventId, MilliSecondsSinceUnixEpoch, OwnedTransactionId, OwnedUserId};
+
+use super::{EventSendState, Profile, TimelineDetails, TimelineItemContent};
+
+/// An item for an event that was created locally and not yet echoed back by
+/// the homeserver.
+#[derive(Debug, Clone)]
+pub struct LocalEventTimelineItem {
+    /// The send state of this local event.
+    pub send_state: EventSendState,
+    /// The transaction ID.
+    pub transaction_id: OwnedTransactionId,
+    /// The sender of the event.
+    pub sender: OwnedUserId,
+    /// The sender's profile of the event.
+    pub sender_profile: TimelineDetails<Profile>,
+    /// The timestamp of the event.
+    pub timestamp: MilliSecondsSinceUnixEpoch,
+    /// The content of the event.
+    pub content: TimelineItemContent,
+}
+
+impl LocalEventTimelineItem {
+    /// Get the event ID of this item.
+    ///
+    /// Will be `Some` if and only if `send_state` is
+    /// `EventSendState::Sent`.
+    pub fn event_id(&self) -> Option<&EventId> {
+        match &self.send_state {
+            EventSendState::Sent { event_id } => Some(event_id),
+            _ => None,
+        }
+    }
+
+    /// Clone the current event item, and update its `send_state`.
+    pub(in crate::room::timeline) fn with_send_state(&self, send_state: EventSendState) -> Self {
+        Self { send_state, ..self.clone() }
+    }
+}
