+++ crates/matrix-sdk/src/room/timeline/event_item/mod.rs
@@ -56,6 +56,10 @@ use ruma::{
 use super::inner::RoomDataProvider;
 use crate::{Error, Result};
 
+mod local;
+
+pub use self::local::LocalEventTimelineItem;
+
 /// An item in the timeline that represents at least one event.
 ///
 /// There is always one main event that gives the `EventTimelineItem` its
@@ -240,41 +244,6 @@ pub enum EventSendState {
         /// The event ID assigned by the server.
         event_id: OwnedEventId,
     },
-}
-
-/// An item for an event that was created locally and not yet echoed back by the
-/// homeserver.
-#[derive(Debug, Clone)]
-pub struct LocalEventTimelineItem {
-    /// The send state of this local event.
-    pub send_state: EventSendState,
-    /// The transaction ID.
-    pub transaction_id: OwnedTransactionId,
-    /// The sender of the event.
-    pub sender: OwnedUserId,
-    /// The sender's profile of the event.
-    pub sender_profile: TimelineDetails<Profile>,
-    /// The timestamp of the event.
-    pub timestamp: MilliSecondsSinceUnixEpoch,
-    /// The content of the event.
-    pub content: TimelineItemContent,
-}
-
-impl LocalEventTimelineItem {
-    /// Get the event ID of this item.
-    ///
-    /// Will be `Some` if and only if `send_state` is `EventSendState::Sent`.
-    pub fn event_id(&self) -> Option<&EventId> {
-        match &self.send_state {
-            EventSendState::Sent { event_id } => Some(event_id),
-            _ => None,
-        }
-    }
-
-    /// Clone the current event item, and update its `send_state`.
-    pub(super) fn with_send_state(&self, send_state: EventSendState) -> Self {
-        Self { send_state, ..self.clone() }
-    }
 }
 
 impl From<LocalEventTimelineItem> for EventTimelineItem {
